@@ -1,25 +1,31 @@
 <script lang="ts">
   import { fade } from 'svelte/transition';
   import { cubicIn } from 'svelte/easing';
-  import { createEventDispatcher, onMount } from 'svelte';
+  import { onMount, type Snippet } from 'svelte';
   import { vibrate } from '$lib/actions/vibrate';
   import RoundClose from 'virtual:icons/ic/round-close';
 
-  export let title: string = '';
-  let scrolled = false;
+  interface Props {
+    title?: string;
+    children?: Snippet;
+    onclose?: () => void;
+  }
+
+  let { title = '', children, onclose }: Props = $props();
+  let scrolled = $state(false);
   let container: HTMLDivElement;
 
-  const dispatch = createEventDispatcher();
-
   onMount(() => {
-    document.onkeydown = (e) => {
+    function keyDownListener(e: KeyboardEvent) {
       if (e.key === 'Escape') {
-        dispatch('close');
+        onclose?.();
       }
-    };
+    }
+
+    document.addEventListener('keydown', keyDownListener);
 
     return () => {
-      document.onkeydown = null;
+      document.removeEventListener('keydown', keyDownListener);
     };
   });
 </script>
@@ -27,14 +33,14 @@
 <div
   transition:fade={{ duration: 200, easing: cubicIn }}
   class="absolute left-0 top-0 z-10 h-full w-full bg-zinc-900/40 backdrop-blur-sm"
-  aria-modal
+  aria-modal="true"
 >
-  <!-- svelte-ignore a11y-no-static-element-interactions -->
-  <!-- svelte-ignore a11y-click-events-have-key-events -->
-  <div class="absolute left-0 top-0 h-full w-full" on:click={() => dispatch('close')} />
+  <!-- svelte-ignore a11y_no_static_element_interactions -->
+  <!-- svelte-ignore a11y_click_events_have_key_events -->
+  <div class="absolute left-0 top-0 h-full w-full" onclick={() => onclose?.()}></div>
   <div
     bind:this={container}
-    on:scroll={() => {
+    onscroll={() => {
       scrolled = container.scrollTop > 0;
     }}
     class="absolute left-0 top-0 m-6 flex h-[calc(100%-3rem)] w-[calc(100%-3rem)] flex-col overflow-auto rounded-md bg-zinc-900 md:m-12 md:h-[calc(100%-6rem)] md:w-[calc(100%-6rem)]"
@@ -48,12 +54,12 @@
       <div
         class="flex items-center justify-center place-self-start rounded-bl-md rounded-tr-md hover:bg-zinc-600/20"
       >
-        <!-- svelte-ignore a11y-autofocus -->
-        <button aria-label="Close modal" autofocus use:vibrate on:click={() => dispatch('close')}>
+        <!-- svelte-ignore a11y_autofocus -->
+        <button aria-label="Close modal" autofocus use:vibrate onclick={() => onclose?.()}>
           <RoundClose class="p-1 text-3xl text-primary/70 transition-colors hover:text-primary" />
         </button>
       </div>
     </div>
-    <slot />
+    {@render children?.()}
   </div>
 </div>
