@@ -13,7 +13,12 @@
   import NavigationElement from './NavigationElement.svelte';
   import { vibrate } from '$lib/actions/vibrate';
 
-  export let user: SignedInUser | null = null;
+  interface Props {
+    user?: SignedInUser | null;
+    onclickedelement?: () => void;
+  }
+
+  let { user = null, onclickedelement }: Props = $props();
 
   type NavigationElementType = {
     href: string;
@@ -69,14 +74,12 @@
       icon: RoundLogin
     }
   ];
-
-  const dispatch = createEventDispatcher();
 </script>
 
 {#if user}
   {#if user.role === 'ADMIN' || user.role === 'OWNER'}
     {#each adminElements as el}
-      <NavigationElement on:clickedelement {...el} />
+      <NavigationElement {onclickedelement} {...el} />
     {/each}
   {/if}
   {#each loggedInElements as el}
@@ -84,7 +87,7 @@
       <form class="px-16 sm:px-0" method="POST" action="/logout" use:enhance>
         <button
           class="flex w-full items-center justify-start gap-2 text-2xl sm:text-base"
-          on:click={() => dispatch('clickedelement')}
+          onclick={() => onclickedelement?.()}
           type="submit"
           use:vibrate
         >
@@ -93,11 +96,11 @@
         </button>
       </form>
     {:else}
-      <NavigationElement on:clickedelement {...el} />
+      <NavigationElement {onclickedelement} {...el} />
     {/if}
   {/each}
 {:else}
   {#each loggedOutElements as el}
-    <NavigationElement on:clickedelement {...el} />
+    <NavigationElement {onclickedelement} {...el} />
   {/each}
 {/if}
